pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "realestate"
version = "0.1.0"
description = "Find real estate deals from interchangeable data sources"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "requests>=2.28",
    "pdfplumber>=0.10",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-dotenv>=1.0",
    "twilio>=9.0",
    "sendgrid>=6.10",
    "python-jose[cryptography]>=3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
realestate = "realestate.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
